feat(sera-memory): extract SemanticMemoryStore + backends from sera-types/sera-db/sera-testing
- Create `sera-memory` crate with lib.rs exposing store.rs, pgvector_store,
  sqlite_store, and in_memory behind feature gates (pgvector/sqlite/testing).
- Add `sera-memory` to workspace members and [workspace.dependencies].
- Delete `sera-types::semantic_memory` module; remove its pub-use re-exports
  from sera-types/src/lib.rs.
- Stub `sera-db::{pgvector_store, sqlite_memory_store}` as one-line
  `pub use sera_memory::*` re-exports so existing callers don't break.
- Stub `sera-testing::semantic_memory` the same way.
- Update all call sites in sera-runtime, sera-workflow, sera-session,
  and sera-gateway to import memory-owned types from `sera_memory` instead
  of `sera_types`. `EmbeddingService` and `memory::SegmentKind` stay in
  sera-types.
- Add `sera-memory` dep to sera-db, sera-testing, sera-runtime,
  sera-workflow, sera-session, and sera-gateway Cargo.toml files.

Validation: `cargo check --workspace` clean, feature matrix green,
`cargo test --workspace` 3582 passed / 13 ignored,
`cargo clippy --workspace -- -D warnings` no issues.
