+++ rust/Cargo.toml
@@ -33,6 +33,7 @@ members = [
   "crates/sera-e2e-harness",
   "crates/sera-mail",
   "crates/sera-eval",
+  "crates/sera-memory",
 ]
 
 [workspace.package]
@@ -188,6 +189,7 @@ url         = "2"
 which       = "8"
 
 # Workspace crates
+sera-memory = { path = "crates/sera-memory" }
 sera-types = { path = "crates/sera-types" }
 sera-telemetry = { path = "crates/sera-telemetry" }
 sera-config = { path = "crates/sera-config" }
