+++ rust/crates/sera-db/Cargo.toml
@@ -9,6 +9,7 @@ integration = []
 sqlite-vec = ["dep:sqlite-vec"]
 
 [dependencies]
+sera-memory = { workspace = true, features = ["pgvector", "sqlite", "sqlite-vec"] }
 sera-types.workspace = true
 sera-errors.workspace = true
 rusqlite.workspace = true
