+++ rust/crates/sera-gateway/Cargo.toml
@@ -14,6 +14,7 @@ path = "src/bin/sera.rs"
 [dependencies]
 clap = { version = "4", features = ["derive"] }
 rusqlite.workspace = true
+sera-memory.workspace = true
 sera-types.workspace = true
 sera-db.workspace = true
 sera-auth.workspace = true
