+++ rust/crates/sera-testing/Cargo.toml
@@ -4,6 +4,7 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
+sera-memory = { workspace = true, features = ["testing"] }
 sera-types.workspace = true
 sera-db.workspace = true
 sera-queue.workspace = true
